+++ src/base/SbDPLine.cpp
@@ -195,10 +195,10 @@ SbDPLine::getClosestPoints(const SbDPLine& line2,
   float c[3], d[3];
 
   for (int i=0; i < 3; i++) {
-    d[i] =
-      D1[i] - D0N[i]*(D0[0]*D1[0] + D0[1]*D1[1] + D0[2]*D1[2]);
-    c[i] =
-      P1[i] - P0[i] + D0N[i]*(D0[0]*P0[0] + D0[1]*P0[1] + D0[2]*P0[2]);
+    d[i] = (float)
+      (D1[i] - D0N[i]*(D0[0]*D1[0] + D0[1]*D1[1] + D0[2]*D1[2]));
+    c[i] = (float)
+      (P1[i] - P0[i] + D0N[i]*(D0[0]*P0[0] + D0[1]*P0[1] + D0[2]*P0[2]));
   }
 
   float t = -(c[0]*d[0]+c[1]*d[1]+c[2]*d[2]) / (d[0]*d[0]+d[1]*d[1]+d[2]*d[2]);
@@ -239,8 +239,8 @@ SbDPLine::getClosestPoint(const SbVec3d& point) const
   //                                        19980815 mortene.
 
 
-  float numerator = (point - this->pos).dot(this->dir);
-  float denumerator = this->dir.length();
+  float numerator = (float) (point - this->pos).dot(this->dir);
+  float denumerator = (float) this->dir.length();
 
 #if COIN_DEBUG
   if(denumerator == 0.0f)
