+++ src/draggers/SoScale1Dragger.cpp
@@ -304,7 +304,7 @@ SoScale1Dragger::drag(void)
   if (motion < 0.0f) motion = 0.0f;
 
   this->setMotionMatrix(this->appendScale(this->getStartMotionMatrix(),
-                                          SbVec3f(fabs(motion), 1.0f, 1.0f),
+                                          SbVec3f((float) fabs(motion), 1.0f, 1.0f),
                                           SbVec3f(0.0f, 0.0f, 0.0f)));
 }
