+++ src/manips/SoClipPlaneManip.cpp
@@ -356,7 +356,7 @@ SoClipPlaneManip::getBoundingBox(SoGetBoundingBoxAction * action)
     action->resetCenter();
   }
   if (numcenters != 0) {
-    action->setCenter(center / numcenters, FALSE);
+    action->setCenter(center / (float) numcenters, FALSE);
   }
 }
